+++ CHANGELOG.md
@@ -2,11 +2,25 @@ Below are the most important changes from each release.
 A more detailed list of changes is available in the corresponding milestones for each release in the Github issue tracker (https://github.com/googlefonts/fontbakery/milestones?state=closed).
 
 
-## 0.7.11 (2019-Aug-16)
+## 0.7.12 (2019-Sep-13)
+### Note-worthy code changes
+  - **[com.google.fonts/check/contour_count]:** ignore PUA codepoints (issue #2612)
+  - **[com.google.fonts/check/contour_count]:** detect glyphs by both glyphnames and codepoints (issue #2612)
+
+### Bugfixes
+  - fix crash on html/markdown reporters by declaring DEBUG log-level (issue #2631)
+
+### Changes to checks
+  - **[com.google.fonts/check/unwanted_tables]:** Add MVAR table (Issue #2599) and improve FAIL message
+
+
+## 0.7.11 (2019-Aug-21)
 ### Note-worthy code changes
   - Adding yet a whole bunch more keywords to log-messages (issue #2558)
+  - Refactored EncodingID/LanguageID classes for Mac & Windows name table entries
 
 ### Bug Fixes
+  - **[com.adobe.fonts/check/family/max_4_fonts_per_family_name]:**  Only run on Win name records (issue #2613)
   - Avoid crash on static fonts by safeguarding the `slnt_axis` condition.
 
 ### New checks